# docs/openapi/packproof-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PackProof API",
    "version": "1.0.0",
    "description": "The contract for PackProof's versioned merchant transaction and public page-declared commerce-handoff API. Transaction and handoff identifiers never confer merchant authority, and API responses intentionally exclude internal persistence, support, secret, and forensic-only fields."
  },
  "servers": [
    {
      "url": "https://YOUR_PACKPROOF_DOMAIN.example",
      "description": "Environment-specific PackProof API origin"
    }
  ],
...
  "paths": {
...
    "/v1/public/integrations/{publishableKey}/handoffs": {
      "post": {
        "operationId": "createPublicCommerceHandoff",
        "tags": ["Commerce Handoffs"],
        "summary": "Create or replay a page-declared passport-draft handoff",
        "description": "Browser-safe endpoint for the PackProof Button. The publishable installation key identifies an integration but is not a secret or merchant authorization. PackProof requires an exact allowlisted HTTPS Origin, binds idempotency to the integration, origin, and canonical request, and caps the result at PAGE_DECLARED trust. This endpoint cannot assert payment, create or bind an external order, or finalize evidence.",
        "security": [],
        "parameters": [
          { "$ref": "#/components/parameters/PublishableKey" },
          { "$ref": "#/components/parameters/Origin" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePublicCommerceHandoffRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/PublicCommerceHandoffCreated" },
          "200": { "$ref": "#/components/responses/PublicCommerceHandoffReplayed" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/transactions": {
      "post": {
        "operationId": "createTransaction",
        "tags": ["Transactions"],
        "summary": "Create or replay a merchant transaction",
        "description": "Creates one immutable PackProof transaction identifier. An exact Idempotency-Key replay returns the original result; a key reused with different input is rejected.",
        "security": [{ "merchantApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTransactionRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/TransactionCreated" },
          "200": { "$ref": "#/components/responses/TransactionReplayed" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "get": {
        "operationId": "listTransactions",
        "tags": ["Transactions"],
        "summary": "List organization transactions",
        "description": "Returns only transactions owned by the authenticated credential's organization, ordered by server creation time and PackProof transaction ID.",
        "security": [{ "merchantApiKey": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/TransactionStatus" }
          },
          {
            "name": "merchantReference",
            "in": "query",
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          },
          {
            "name": "createdAfter",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "createdBefore",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string", "minLength": 8, "maxLength": 2000 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 }
          }
        ],
        "responses": {
          "200": {
            "description": "An organization-isolated transaction page.",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/XRequestId" },
              "RateLimit-Limit": { "$ref": "#/components/headers/RateLimitLimit" },
              "RateLimit-Remaining": { "$ref": "#/components/headers/RateLimitRemaining" },
              "RateLimit-Reset": { "$ref": "#/components/headers/RateLimitReset" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListTransactionsResponse" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/transactions/{transactionId}": {
      "get": {
        "operationId": "getTransaction",
        "tags": ["Transactions"],
        "summary": "Retrieve one organization transaction",
        "security": [{ "merchantApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TransactionId" }
        ],
        "responses": {
          "200": {
            "description": "The authorized merchant transaction.",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/XRequestId" },
              "RateLimit-Limit": { "$ref": "#/components/headers/RateLimitLimit" },
              "RateLimit-Remaining": { "$ref": "#/components/headers/RateLimitRemaining" },
              "RateLimit-Reset": { "$ref": "#/components/headers/RateLimitReset" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GetTransactionResponse" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/transactions/{transactionId}/participant-invitations": {
      "post": {
        "operationId": "createParticipantInvitation",
        "tags": ["Participant Claims"],
        "summary": "Create or replay a participant invitation",
        "description": "Issues a short-lived, one-time claim capability for a participant reference already declared on this merchant transaction. The external reference is a non-authorizing merchant label: only a successful claim made by an authenticated PackProof user binds an actor to the role. The plaintext token is returned only in this response and is stored by PackProof only as a digest.",
        "security": [{ "merchantApiKey": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TransactionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateParticipantInvitationRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ParticipantInvitationCreated" },
          "200": { "$ref": "#/components/responses/ParticipantInvitationReplayed" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/participant-claims": {
      "post": {
        "operationId": "claimParticipantInvitation",
        "tags": ["Participant Claims"],
        "summary": "Claim a declared transaction role",
        "description": "Binds the authenticated PackProof user to the invitation's transaction role after validating the one-time claim token. Firebase App Check augments, but never replaces, Firebase user identity. Possession of a claim link alone provides no transaction read authority.",
        "security": [{ "firebaseUserBearer": [], "firebaseAppCheck": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClaimParticipantRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ParticipantClaimed" },
          "200": { "$ref": "#/components/responses/ParticipantClaimReplayed" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "merchantApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "pp_{environment}_{credentialId}.{secret}",
        "description": "Environment-scoped merchant credential. The secret is shown once and is never stored in plaintext by PackProof."
      },
      "firebaseUserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "A current Firebase ID token identifying the PackProof user. Claim and redemption links do not replace user authentication."
      },
      "firebaseAppCheck": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Firebase-AppCheck",
        "description": "A valid Firebase App Check token. It supplies application context and augments, but does not replace, user identity or resource authorization."
      }
    },
    "parameters": {
...
      "PublishableKey": {
        "name": "publishableKey",
        "in": "path",
        "required": true,
        "description": "Public PackProof Button installation identifier. It is safe to place in browser code and grants no merchant or order authority.",
        "schema": { "type": "string", "pattern": "^pp_pub_(sandbox|live)_[A-Za-z0-9_-]{20,80}$" }
      },
...
    },
...
    "responses": {
...
      "InvalidRequest": { "$ref": "#/components/responses/Error400" },
...
      "Forbidden": { "$ref": "#/components/responses/Error403" },
...
      "Conflict": { "$ref": "#/components/responses/Error409" },
...
    }
  }
}